Add Publishing API docs, App docs, DocumentTypes/Supertypes to search
- Moves this config into a new class 'ProxyPages' so that it can be tested
- Adds a 'title' frontmatter property so that it is indexed by middleman-search

This makes the search far more comprehensive and useful, although it does
slow the initial indexing down a bit, so the 'Loading search index'
message will stick around slightly longer. I think this is worth it for
the improved relevancy of search results, and if it becomes a problem we
can investigate a server-side search/index solution, or make better use
of caching.

Unfortunately I could not get the govuk_schema_names indexed in the search,
as this practically breaks the indexing altogether. I haven't yet found
a way of indexing just by title and not content.

Trello: https://trello.com/c/lVTGSjCA/142-imported-dev-docs-dont-appear-in-search